+++ CHANGELOG.md
@@ -6,6 +6,7 @@ If you are upgrading from v3.x.x,  please see ['UPGRADING.md'](UPGRADING.md)
 
 ### Fixed
 - fix vulnerabilities from `npm audit`
+- fix sdk usage from CDN
 - fix silent sign in flow
 - fix device flow
 - fix enrollment flow
@@ -14,7 +15,6 @@ If you are upgrading from v3.x.x,  please see ['UPGRADING.md'](UPGRADING.md)
 - fix updateProfileImage() function
 - fix getUserActivities() function
 - fix error if query parameter are not included in userCheckExists() function
-- fix sdk usage from CDN
 - fix missing logo on npmjs
 
 ### Added
@@ -23,9 +23,9 @@ If you are upgrading from v3.x.x,  please see ['UPGRADING.md'](UPGRADING.md)
 - add initiateDeviceCode() & offlineTokenCheck() functions to token service
 
 ### Changed
-- **Breaking** Authentication Module can't be access publicly anymore, instead Webauth should be used to access Authentication functions.
+- **Breaking** authentication module can't be access publicly anymore, instead WebAuth should be used to access authentication functions.
 - **Breaking** popup & silent authentication functions is directly implemented in WebAuth instead of using mode.
-- **Breaking** silentSignIn is now returning Promise
+- **Breaking** silentSignIn is now returning promise
 - **Breaking** getLoginURL() function return promise instead of string
 - **Breaking** access_token option should be provided to revokeClaimConsent() function in consent service
 - **Breaking** use function parameter instead of cidaas configuration file: `cidaas_version` to handle resetPassword